requirements.txt
pandas
Python-dotenv
requests
lxml
html5lib
beautifulsoup4